Bugfixing
1. Fixed blender plane asset (wrong rotation)
2. Replaced hardcoded quad with plane asset.
3. A bit of integer overflow protection + enforcing uint32_t everywhere as in original dx12 api.
4. Fixed bounding box for flat objects like a plane.

+++ Engine/renderer/command_list.h
@@ -31,7 +31,7 @@ namespace engine
     void clear_render_target(const ftexture_resource* rtv) const;
     void clear_depth_stencil(const ftexture_resource* dsv) const;
 
-    void upload_buffer_resource(uint64_t buffer_size, const void* in_buffer, ComPtr<ID3D12Resource>& out_upload_intermediate, ComPtr<ID3D12Resource>& out_gpu_resource) const;
+    void upload_buffer_resource(uint32_t buffer_size, const void* in_buffer, ComPtr<ID3D12Resource>& out_upload_intermediate, ComPtr<ID3D12Resource>& out_gpu_resource) const;
     void upload_vertex_buffer(astatic_mesh* mesh, const char* name) const;
     void upload_index_buffer(astatic_mesh* mesh, const char* name) const;
     void upload_texture(atexture* texture_asset) const;
